+++ drivers/hwinfo/hwinfo_litex.c
@@ -12,10 +12,10 @@
 #include <zephyr/device.h>
 #include <zephyr/sys/util.h>
 
-ssize_t z_impl_hwinfo_get_device_id(uint8_t *buffer, size_t length)
+int z_impl_hwinfo_get_device_id(uint8_t *buffer, size_t length)
 {
 	uint32_t addr = DT_INST_REG_ADDR(0);
-	ssize_t end = MIN(length, DT_INST_REG_ADDR(0) / 4 *
+	int end = MIN(length, DT_INST_REG_ADDR(0) / 4 *
 			CONFIG_LITEX_CSR_DATA_WIDTH / 8);
 	for (int i = 0; i < end; i++) {
 #if CONFIG_LITEX_CSR_DATA_WIDTH == 8
