+++ drivers/hwinfo/hwinfo_weak_impl.c
@@ -6,7 +6,7 @@
 
 #include <zephyr/drivers/hwinfo.h>
 
-ssize_t __weak z_impl_hwinfo_get_device_id(uint8_t *buffer, size_t length)
+int __weak z_impl_hwinfo_get_device_id(uint8_t *buffer, size_t length)
 {
 	return -ENOSYS;
 }
